pulled code from main branch and merged. refactored code in SignInButton to reflect changes. Reverted framer-motion to earlier version.
